backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
torch==2.6.0
torchvision==0.21.0
transformers==4.46.0
opencv-python==4.10.0.84
pillow==11.0.0
numpy==2.1.0
python-dotenv==1.0.1
ffmpeg-python==0.2.0
requests==2.31.0
yt-dlp==2024.8.6
PyPDF2==3.0.1
python-docx==1.1.0
validators==0.22.0
langdetect==1.0.9
deep-translator==1.11.4
sentencepiece==0.1.99